refactor: drop temporary re-exports from parser.ts
Point every importer at the concrete module that defines each symbol
(substitution.ts, pbScript.ts, envFiles.ts, workspaceTree.ts) and delete
the transitional re-export block from parser.ts. parser.ts now keeps only
the .http parse/serialize core it defines itself, removing the back edge
where parser.ts re-exported workspaceTree.ts while workspaceTree.ts
imports parser.ts.

Mechanical import updates only - no logic changes.
