+++ src/lib/importIO.ts
@@ -9,7 +9,8 @@ import { writeTextFile, mkdir } from '@tauri-apps/plugin-fs';
 import { join, basename, dirname } from '@tauri-apps/api/path';
 import { workspace, envFile, activeEnvironment, addToast } from './stores';
 import { errorMessage } from './errors';
-import { buildWorkspaceTree, ensureSharedEnvironment } from './parser';
+import { ensureSharedEnvironment } from './envFiles';
+import { buildWorkspaceTree } from './workspaceTree';
 import { scanForHttpFiles, safeJoinPath } from './workspaceIO';
 import { importPostmanCollection } from './postman';
 import { importInsomniaExport } from './insomnia';
